+++ client/src/hooks/index.ts
@@ -7,5 +7,6 @@ export { default as useLocalize } from './useLocalize';
 export { default as useMediaQuery } from './useMediaQuery';
 export { default as useSetOptions } from './useSetOptions';
 export { default as useGenerations } from './useGenerations';
+export { default as useScrollToRef } from './useScrollToRef';
 export { default as useServerStream } from './useServerStream';
 export { default as useMessageHandler } from './useMessageHandler';
