+++ client/src/routes/Search.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useRecoilState, useRecoilValue } from 'recoil';
 
-import Messages from '~/components/Messages';
+import Messages from '~/components/Messages/Messages';
 import TextChat from '~/components/Input/TextChat';
 
 import store from '~/store';
